refactor(roslyn): extract shared UsingDirectiveSynthesizer
The using-directive synthesis cluster — BuildUsingDirectives plus its six
helpers (PreserveSpecialAndRequiredSourceUsings, AddMissingRequiredUsingDirectives,
SortUsingDirectives, GetUsingNamespace, IsSystemUsingDirective,
IsSpecialUsingDirective) — was byte-identical in InterfaceExtractionService and
CrossProjectRefactoringService. The source doc-comment literally said
"Mirrors InterfaceExtractionService.BuildUsingDirectives"; the two copies could
drift independently.

Hoist the cluster into RoslynMcp.Roslyn.Helpers.UsingDirectiveSynthesizer
(internal static, mirroring the TypeNamespaceWalker/DiffGenerator helper pattern).
Both services now delegate to it; behavior is unchanged. Add a focused regression
test on the shared helper (preserve aliases/static, drop unrequired plain usings,
add missing required, System-first sort) — testing it once covers both entry
points since they share the code.

Closes using-directive-synthesis-dedup.
