+++ ai_docs/backlog.md
@@ -3,7 +3,7 @@
 <!-- purpose: Open work only; contract for agents syncing backlog on ship. -->
 <!-- scope: in-repo -->
 
-**updated_at: 2026-05-30T16:53:21Z**
+**updated_at: 2026-05-30T20:02:00Z**
 
 ## Agent contract
 
@@ -51,7 +51,6 @@
 | id | pri | deps | do |
 |----|-----|------|-----|
 | `promotion-scorecard-execution-batch` | Medium | none | Promotion scorecard (2026-05-08) lists 25 promote-ready experimental tools against server 1.34.2; current is 2.2.2 (18 days drift). Re-run scorecard against the 2.2.2 surface first; resolve the duplicate-source-of-truth between `audit-reports/_latest-promotion-scorecard.json` (serverVersion 1.38.1) and `ai_docs/audit-reports/_latest-promotion-scorecard.json` (serverVersion 1.34.2) by picking one canonical location and deleting the other; then ship tier promotions in bounded batches via the `/promote-tier` skill. Anchors: `audit-reports/_latest-promotion-scorecard.json`, `ai_docs/audit-reports/_latest-promotion-scorecard.json`, `src/RoslynMcp.Host.Stdio/Catalog/ServerSurfaceCatalog.*.cs`. Clusters: brainstorm BRAIN-007 (stable-surface promotion scorecard system). Evidence: scorecard `summary.promoteReady=25`; no recent commits matching "promote"/"tier" in last 6 weeks. Source: 2026-05-26 discovery-sweep work-search + brainstorm. **Top-5 2026-05-27 audit: sweep-shaped — multi-file source-of-truth decision + bounded-batch execution; recommend `/backlog-sweep:prepare` rather than top-5 remediation.** |
-| `using-directive-synthesis-dedup` | Medium | none | Byte-identical using-directive synthesis cluster duplicated across two refactoring services — `BuildUsingDirectives` (15 LOC) + `PreserveSpecialAndRequiredSourceUsings` (24 LOC) plus the `AddMissingRequiredUsingDirectives`/`SortUsingDirectives`/`IsSpecialUsingDirective`/`GetUsingNamespace` helpers they call; the source doc-comment literally says "Mirrors InterfaceExtractionService.BuildUsingDirectives". Cross-service copies drift independently. Extract the cluster into a shared static helper (e.g. `Helpers/UsingDirectiveSynthesizer.cs`) and call it from both. Anchors: `src/RoslynMcp.Roslyn/Services/CrossProjectRefactoringService.cs:874`, `src/RoslynMcp.Roslyn/Services/InterfaceExtractionService.cs:374`. Regression test shape: existing interface-extraction + cross-project move tests stay green; one test asserts both entry points produce identical using-lists for the same input. Evidence: `find_duplicated_code` similarity=1.0 (semantic, workspace-loaded). Source: 2026-05-28 discovery-sweep refactor. [type: refactor] |
 | `audit-21-analyzer-load-decision` | Medium | product decision (does the product require full IDE/CA analyzer parity?) | AUDIT-21: IDE/CA analyzers are not host-injected into `MSBuildWorkspace`, so Roslyn-MCP diagnostics can differ from Visual Studio / Rider. Real unshipped gap with a detailed Draft plan (`ai_docs/plans/audit-21-analyzers/plan.md`, status active/dormant, owner TBD) but no prior backlog tracking; `docs/parity-gap-implementation-plan.md` gates it on "fix only if product requires full analyzer load". Decide: (a) execute via `/backlog-sweep:prepare` against the plan (sweep-shaped, ~12K-token plan), or (b) re-status the plan to `superseded`/parked with the explicit product trigger so a dormant Draft stops sitting in the active router. Either way, fix the plan's §13 closure criterion which still cites the removed row `compile-check-vs-analyzers-doc`. Anchors: `ai_docs/plans/audit-21-analyzers/plan.md`, `ai_docs/architecture.md:83`, `docs/parity-gap-implementation-plan.md:76`, `src/RoslynMcp.Roslyn/Services/CompilationCache.cs`. Evidence: no host-analyzer ship commits in `git log --all`; `IHostAnalyzerProvider` absent from `src/`; distinct from brainstorm BRAIN-010 (parity transparency report). Source: 2026-05-28 discovery-sweep work-search + doc-audit Known Gap. |
 | `compilation-cache-adoption-read-side` | Medium | none | Read-side analysis services call raw `project.GetCompilationAsync(ct)` against the live workspace solution instead of the singleton `ICompilationCache`, forgoing cross-tool/cross-request compilation sharing + analyzer-bound caching — only 4 of ~24 `GetCompilationAsync` call sites use the cache today (`DiagnosticService`, `DiRegistrationService`, `NamespaceDependencyService`, `UnusedCodeAnalyzer`). First bounded batch — 3 verified-clean sites (instance method, `workspaceId` in scope, `GetCurrentSolution` live, single in-loop call → one-line swap to `_compilationCache.GetCompilationAsync(workspaceId, project, ct)` + ctor field): `CouplingAnalysisService` (hottest — ~12.8s p50 on a 7-proj/571-doc solution per its own inline note), `ExceptionFlowService`, `AnalyzerInfoService`. Note: benefit is narrower than "rebuild→cached" because Roslyn already memoizes a `Project`'s compilation internally; the real wins are guaranteed cross-call sharing under GC pressure, analyzer-bound compilations, and concurrency in-flight dedup. Anchors: `src/RoslynMcp.Roslyn/Services/CouplingAnalysisService.cs:50`, `src/RoslynMcp.Roslyn/Services/ExceptionFlowService.cs:60`, `src/RoslynMcp.Roslyn/Services/AnalyzerInfoService.cs:34`, `src/RoslynMcp.Roslyn/Services/CompilationCache.cs`. Regression test shape: existing coupling/exception-flow/analyzer-list tests stay green; one test asserts two successive read-tool calls at an unchanged workspace version resolve the reference-equal cached `Compilation` for a given project. Follow-on surface (NOT this row's scope — split into child rows at planning): (a) static-helper read-side sites needing `workspaceId`+cache threaded through `Solution`-only signatures (`TypeConsumersService:108`, `TestReferenceMapService:150/178`, `CodePatternAnalyzer:229`, `SymbolSearchService:117`, `ReferenceService:231`; secondary-path `ImpactSweepService`, `MutationAnalysisService`, `SymbolRelationshipService:427`); (b) pre-edit live fetches in preview services (subagent-triaged, per-site confirm) `BuildService:149`, `FixAllService:306`, `BulkRefactoringService:166/554/574`, `CrossProjectRefactoringService:376`, `InterfaceExtractionService:494`, `ScaffoldingService.*`; (c) HAZARD — high-leverage `solution`-parameterized helpers `SymbolResolver.ResolveByMetadataNameAsync:285`/`FindClosestMatchesAsync:125` + `SymbolHandleSerializer` are called with a FORKED solution by `CrossProjectRefactoringService:226`, so routing them through the version-keyed cache must gate on `solution == GetCurrentSolution(workspaceId)` or it serves a stale compilation. Confirmed forked exclusions (must NOT cache): `InterfaceExtractionService:515`, `RefactoringService:415`, `TypeMoveService:217`. Evidence: 4/24 adoption; `CouplingAnalysisService.cs:61-67` inline p50 note; `ICompilationCache` version-keyed live-solution-only contract. Source: 2026-05-29 memoization audit — first batch (3 sites) personally verified, remainder subagent-triaged. [type: refactor] |
